backend/requirements.txt
# Core FastAPI and Web Framework
fastapi
uvicorn
python-multipart

# AI and LangChain (Essential only)
langchain
langchain-community
langchain-core
langchain-groq

# Database (SQLite only for simplicity)
sqlalchemy

# Environment and Configuration
python-dotenv

# Basic Security
python-jose[cryptography]
passlib[bcrypt]

# Utilities
python-slugify
